requirements.txt
# 법제처 입법자료 변환기 필수 패키지

# 문서 처리
python-docx==1.1.0

# HTML 파싱
beautifulsoup4==4.12.3
lxml==5.1.0

# 웹 UI
streamlit==1.31.0
